electron build works almost

+++ plugins/electron/compile.js
@@ -16,6 +16,13 @@ export default function() {
 
     options.output.path = buildDir;
     let runner = webpack(options);
+    if (!this.nuxt.dev) {
+      runner.run(() => {
+        debugger;
+      });
+      return;
+    }
+
     runner.watch(nuxt.options.watch, (err) => {
       if (electronProc) {
         electronProc.kill("SIGINT");
